@@ -1,11 +1,10 @@
 buildscript {
     repositories {
-        maven { url = 'https://files.minecraftforge.net/maven' }
-        jcenter()
+        maven { url = 'https://maven.minecraftforge.net' }
         mavenCentral()
     }
     dependencies {
-        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '3.+', changing: true
+        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '5.1.+', changing: true
     }
 }
 apply plugin: 'net.minecraftforge.gradle'
@@ -17,19 +16,20 @@ version = project.mod_version
 group = 'net.ommina.wallpapercraft' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "${name}-${minecraft_version}"
 
-sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8' // Need this here so eclipse task generates correctly.
+java.toolchain.languageVersion = JavaLanguageVersion.of(16)
+println('Java: ' + System.getProperty('java.version') + ' JVM: ' + System.getProperty('java.vm.version') + '(' + System.getProperty('java.vendor') + ') Arch: ' + System.getProperty('os.arch'))
 
-processResources {
-    from(sourceSets.main.resources.srcDirs) {
-        include 'META-INF/mods.toml'
+//processResources {
+//    from(sourceSets.main.resources.srcDirs) {
+//        include 'META-INF/mods.toml'
 
-        expand 'mod_version': mod_version, 'forge_version_min': forge_version_min, 'forge_version_max': forge_version_max, 'minecraft_version_min': minecraft_version_min, 'minecraft_version_max': minecraft_version_max
-    }
+//        expand 'mod_version': mod_version, 'forge_version_min': forge_version_min, 'forge_version_max': forge_version_max, 'minecraft_version_min': minecraft_version_min, 'minecraft_version_max': minecraft_version_max
+//    }
 
-    from(sourceSets.main.resources.srcDirs) {
-        exclude 'META-INF/mods.toml'
-    }
-}
+//from(sourceSets.main.resources.srcDirs) {
+//    exclude 'META-INF/mods.toml'
+//}
+//}
 
 minecraft {
 
@@ -92,6 +92,8 @@ minecraft {
     }
 }
 
+sourceSets.main.resources { srcDir 'src/generated/resources' }
+
 task sourcesJar(type: Jar) {
     classifier = "sources"
     from sourceSets.main.allJava
@@ -114,6 +116,7 @@ repositories {
     }
 }
 
+
 dependencies {
 
     minecraft([
@@ -123,9 +126,10 @@ dependencies {
     ])
 
     // compile against the JEI API but do not include it at runtime
-    compileOnly fg.deobf("mezz.jei:jei-1.16.2:7.3.1.23:api")
+    // deobfProvided "mezz.jei:jei_${mc_version}:${jei_version}:api"
+    compileOnly fg.deobf("mezz.jei:jei-${minecraft_version}:${jei_version}:api")
     // at runtime, use the full JEI jar
-    runtimeOnly fg.deobf("mezz.jei:jei-1.16.2:7.3.1.23")
+    runtimeOnly fg.deobf("mezz.jei:jei-${minecraft_version}:${jei_version}")
 
 }
 
@@ -144,10 +148,13 @@ jar {
     }
 }
 
+jar.finalizedBy('reobfJar')
+
 // Example configuration to allow publishing using the maven-publish task
 // we define a custom artifact that is sourced from the reobfJar output task
 // and then declare that to be published
 // Note you'll need to add a repository here
+
 def reobfFile = file("$buildDir/reobfJar/output.jar")
 def reobfArtifact = artifacts.add('default', reobfFile) {
     type 'jar'
