@@ -1,16 +1,15 @@
 package net.ommina.wallpapercraft.blocks;
 
-import net.minecraft.block.AbstractGlassBlock;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.SoundType;
-import net.minecraft.block.material.Material;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IWorldReader;
-import net.minecraft.world.World;
-import net.minecraftforge.common.ToolType;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.AbstractGlassBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Material;
 import net.ommina.wallpapercraft.items.ModItems;
 import net.ommina.wallpapercraft.items.PressColour;
 import net.ommina.wallpapercraft.items.PressVariant;
@@ -26,16 +25,14 @@ public class DecorativeBlockGlass extends AbstractGlassBlock implements IDecorat
     private final String colour;
     private final String suffix;
 
-    public DecorativeBlockGlass( final String pattern, final String colour, final int suffix, final Material material, final ToolType toolType, final SoundType soundType, final float hardness, final int light ) {
+    public DecorativeBlockGlass( final String pattern, final String colour, final int suffix, final Material material, final SoundType soundType, final float hardness, final int light ) {
 
-        super( Block.Properties.create( material )
+        super( Block.Properties.of( material )
              .sound( soundType )
-             .harvestTool( toolType )
-             .harvestLevel( 0 )
-             .hardnessAndResistance( hardness )
-             .notSolid()
-             .setRequiresTool()
-             .setLightLevel( ( l ) -> light )
+             .strength( hardness )
+             .noOcclusion()
+             .requiresCorrectToolForDrops()
+             .lightLevel( ( l ) -> light )
         );
 
         this.pattern = pattern;
@@ -45,6 +42,11 @@ public class DecorativeBlockGlass extends AbstractGlassBlock implements IDecorat
     }
 
     //region Overrides
+    @Override
+    public void attack( final BlockState state, final Level world, final BlockPos pos, final Player player ) {
+        IDecorativeBlock.super.onBlockClicked( state, world, pos, player );
+    }
+
     @Override
     public String getPostfix() {
         return POSTFIX;
@@ -67,28 +69,22 @@ public class DecorativeBlockGlass extends AbstractGlassBlock implements IDecorat
     }
 
     @Override
-    public SoundType getSoundType( final BlockState state, final IWorldReader world, final BlockPos pos, @Nullable final Entity entity ) {
+    public SoundType getSoundType( final BlockState state, final LevelReader world, final BlockPos pos, @Nullable final Entity entity ) {
 
-        if ( !(entity instanceof PlayerEntity) )
+        if ( !(entity instanceof Player) )
             return SoundType.GLASS;
 
-        final PlayerEntity player = (PlayerEntity) entity;
+        final Player player = (Player) entity;
 
-        if ( player.getHeldItemMainhand().isEmpty() )
+        if ( player.getMainHandItem().isEmpty() )
             return SoundType.GLASS;
 
-        if ( player.getHeldItemMainhand().getItem() == ModItems.PAINTBRUSH || player.getHeldItemMainhand().getItem() instanceof PressColour || player.getHeldItemMainhand().getItem() instanceof PressVariant )
+        if ( player.getMainHandItem().getItem() == ModItems.PAINTBRUSH || player.getMainHandItem().getItem() instanceof PressColour || player.getMainHandItem().getItem() instanceof PressVariant )
             return ModSoundType.BLOCK_CHANGE;
 
         return SoundType.STONE;
 
     }
-
-    @Override
-    public void onBlockClicked( final BlockState state, final World world, final BlockPos pos, final PlayerEntity player ) {
-        IDecorativeBlock.super.onBlockClicked( state, world, pos, player );
-    }
-
     //endregion Overrides
 
 }
