@@ -1,11 +1,11 @@
 package net.ommina.wallpapercraft.blocks;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
 import net.ommina.wallpapercraft.items.ModItems;
 import net.ommina.wallpapercraft.items.PressColour;
 import net.ommina.wallpapercraft.items.PressVariant;
@@ -22,22 +22,22 @@ public interface IDecorativeBlock {
 
     String getSuffix();
 
-    default void onBlockClicked( final BlockState state, final World world, final BlockPos pos, final PlayerEntity player ) {
+    default void onBlockClicked( final BlockState state, final Level world, final BlockPos pos, final Player player ) {
 
         Block block = Blocks.AIR;
 
-        if ( player.getHeldItemMainhand().getItem() == ModItems.PAINTBRUSH )
+        if ( player.getMainHandItem().getItem() == ModItems.PAINTBRUSH )
             block = InWorldHelper.getIncrementedBlockColour( this );
-        else if ( player.getHeldItemMainhand().getItem() instanceof PressColour )
-            block = InWorldHelper.getBlockFromColourPress( this, (PressColour) player.getHeldItemMainhand().getItem() );
-        else if ( player.getHeldItemMainhand().getItem() instanceof PressVariant )
-            block = InWorldHelper.getBlockFromVariantPress( this, (PressVariant) player.getHeldItemMainhand().getItem() );
+        else if ( player.getMainHandItem().getItem() instanceof PressColour )
+            block = InWorldHelper.getBlockFromColourPress( this, (PressColour) player.getMainHandItem().getItem() );
+        else if ( player.getMainHandItem().getItem() instanceof PressVariant )
+            block = InWorldHelper.getBlockFromVariantPress( this, (PressVariant) player.getMainHandItem().getItem() );
 
         if ( block == Blocks.AIR )
             return;
 
-        if ( !world.isRemote )
-            world.setBlockState( pos, block.getDefaultState(), 3 );
+        if ( !world.isClientSide )
+            world.setBlock( pos, block.defaultBlockState(), 3 );
 
     }
 
