@@ -1,8 +1,8 @@
 package net.ommina.wallpapercraft.client;
 
 import net.minecraft.client.Minecraft;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.world.World;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
 import net.ommina.wallpapercraft.IProxy;
 
 public class ClientProxy implements IProxy {
@@ -12,12 +12,12 @@ public class ClientProxy implements IProxy {
     }
 
     @Override
-    public World getClientWorld() {
-        return Minecraft.getInstance().world;
+    public Level getClientWorld() {
+        return Minecraft.getInstance().level;
     }
 
     @Override
-    public PlayerEntity getClientPlayer() {
+    public Player getClientPlayer() {
         return Minecraft.getInstance().player;
     }
 
