@@ -1,7 +1,9 @@
 package net.ommina.wallpapercraft.items;
 
-import net.minecraft.block.Block;
-import net.minecraft.item.BlockItem;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.item.BlockItem;
+
+import net.minecraft.world.item.Item.Properties;
 
 public class DecorativeItem extends BlockItem {
     public DecorativeItem( final Block blockIn, final Properties builder ) {
