@@ -1,7 +1,7 @@
 package net.ommina.wallpapercraft.server;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.world.World;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.ommina.wallpapercraft.IProxy;
@@ -14,12 +14,12 @@ public final class ServerProxy implements IProxy {
     }
 
     @Override
-    public World getClientWorld() {
+    public Level getClientWorld() {
         throw new IllegalStateException( "Only run this on the client!" );
     }
 
     @Override
-    public PlayerEntity getClientPlayer() {
+    public Player getClientPlayer() {
         throw new IllegalStateException( "Only run this on the client!" );
     }
 
