@@ -1,14 +1,14 @@
 package net.ommina.wallpapercraft.sounds;
 
-import net.minecraft.block.SoundType;
-import net.minecraft.util.SoundEvent;
-import net.minecraft.util.SoundEvents;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.ommina.wallpapercraft.DeferredRegistries;
 
 public class ModSoundType extends SoundType {
 
-    public static ModSoundType BLOCK_CHANGE = new ModSoundType( 1.0F, 1.0F, SoundEvents.BLOCK_STONE_BREAK, SoundEvents.BLOCK_STONE_STEP, SoundEvents.BLOCK_STONE_PLACE, DeferredRegistries.BLOCK_CHANGE_IN_WORLD, SoundEvents.BLOCK_STONE_FALL );
+    public static ModSoundType BLOCK_CHANGE = new ModSoundType( 1.0F, 1.0F, SoundEvents.STONE_BREAK, SoundEvents.STONE_STEP, SoundEvents.STONE_PLACE, DeferredRegistries.BLOCK_CHANGE_IN_WORLD, SoundEvents.STONE_FALL );
 
     private RegistryObject<SoundEvent> hitSound;
 
