@@ -1,9 +1,9 @@
 package net.ommina.wallpapercraft.util;
 
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.resources.I18n;
-import net.minecraft.client.resources.Language;
-import net.minecraft.client.resources.LanguageManager;
+import net.minecraft.client.resources.language.I18n;
+import net.minecraft.client.resources.language.LanguageInfo;
+import net.minecraft.client.resources.language.LanguageManager;
 
 import java.util.Locale;
 
@@ -13,11 +13,11 @@ public class Translator {
     }
 
     public static String translateToLocal(String key) {
-        return I18n.format(key);
+        return I18n.get(key);
     }
 
     public static String translateToLocalFormatted(String key, Object... format) {
-        return I18n.format(key, format);
+        return I18n.get(key, format);
     }
 
     public static String toLowercaseWithLocale(String string) {
@@ -30,7 +30,7 @@ public class Translator {
         if (minecraft != null) {
             LanguageManager languageManager = minecraft.getLanguageManager();
             if (languageManager != null) {
-                Language currentLanguage = languageManager.getCurrentLanguage();
+                LanguageInfo currentLanguage = languageManager.getSelected();
                 if (currentLanguage != null) {
                     return currentLanguage.getJavaLocale();
                 }
